+++ Source/Android/jni/MainAndroid.cpp
@@ -37,7 +37,6 @@
 
 #include "DiscIO/Enums.h"
 #include "DiscIO/Volume.h"
-#include "DiscIO/VolumeCreator.h"
 
 #include "UICommon/UICommon.h"
