+++ Source/Core/Core/ConfigManager.cpp
@@ -36,7 +36,6 @@
 #include "DiscIO/Enums.h"
 #include "DiscIO/NANDContentLoader.h"
 #include "DiscIO/Volume.h"
-#include "DiscIO/VolumeCreator.h"
 
 SConfig* SConfig::m_Instance;
