+++ Source/Core/DiscIO/DiscIO.vcxproj.filters
@@ -54,9 +54,6 @@
     <ClCompile Include="WbfsBlob.cpp">
       <Filter>Volume\Blob</Filter>
     </ClCompile>
-    <ClCompile Include="VolumeCreator.cpp">
-      <Filter>Volume</Filter>
-    </ClCompile>
     <ClCompile Include="VolumeDirectory.cpp">
       <Filter>Volume</Filter>
     </ClCompile>
@@ -117,9 +114,6 @@
       <Filter>Volume\Blob</Filter>
     </ClInclude>
     <ClInclude Include="Volume.h">
-      <Filter>Volume</Filter>
-    </ClInclude>
-    <ClInclude Include="VolumeCreator.h">
       <Filter>Volume</Filter>
     </ClInclude>
     <ClInclude Include="VolumeDirectory.h">
