+++ Source/Core/DiscIO/DiscScrubber.cpp
@@ -16,7 +16,6 @@
 #include "DiscIO/DiscScrubber.h"
 #include "DiscIO/Filesystem.h"
 #include "DiscIO/Volume.h"
-#include "DiscIO/VolumeCreator.h"
 
 namespace DiscIO
 {
