+++ Source/Core/DiscIO/VolumeWiiCrypted.cpp
@@ -25,7 +25,6 @@
 #include "DiscIO/Enums.h"
 #include "DiscIO/Filesystem.h"
 #include "DiscIO/Volume.h"
-#include "DiscIO/VolumeCreator.h"
 
 namespace DiscIO
 {
