+++ Source/Core/DolphinWX/ISOProperties/FilesystemPanel.cpp
@@ -24,7 +24,6 @@
 #include "DiscIO/Enums.h"
 #include "DiscIO/Filesystem.h"
 #include "DiscIO/Volume.h"
-#include "DiscIO/VolumeCreator.h"
 #include "DolphinWX/ISOFile.h"
 #include "DolphinWX/WxUtils.h"
